rename AccountReader to Account, give a separate module
